Iterator over all the enabled options
This adds a new function that returns an iterator over all the enabled flags.

This implementation takes into account combined flags like the ones described in: https://github.com/bitflags/bitflags/pull/204#issuecomment-950304444

It uses an `impl Iterator` as return type because using an specific Iter type implies using const generics which are not supported in earlier versions of rust and adding a const to the trait in order to express the correct dimension of the array as can be seen in the last solution proposed in #204.

This superseeds #204